chore(#71): refactor app name on main.dart file
