+++ lib/main.dart
@@ -4,6 +4,7 @@ import 'package:parental_control/services/app_usage_service.dart';
 import 'package:parental_control/services/auth.dart';
 import 'package:parental_control/services/geo_locator_service.dart';
 import 'package:parental_control/theme/theme.dart';
+import 'package:parental_control/utils/app_strings.dart';
 import 'package:provider/provider.dart';
 
 import 'app/config/screencontroller_config.dart';
@@ -43,7 +44,7 @@ class _MyAppState extends State<MyApp> {
 
     return MaterialApp(
       debugShowCheckedModeBanner: false,
-      title: 'Parental Control',
+      title: Strings.appName,
       theme: AppTheme.lightTheme,
       darkTheme: AppTheme.darkTheme,
       home: FutureBuilder(
