+++ lib/utils/app_strings.dart
@@ -1,0 +1,4 @@
+class Strings{
+
+  static const String appName= "Time's Up";
+}
